requirements.txt
numpy
python-dotenv
PyPDF2
streamlit
requests
pandas
beautifulsoup4
langchain
openai
tiktoken